Joined Login and register in one page, and made it the homePublic page
