+++ src/App.tsx
@@ -4,8 +4,9 @@ import { Route, Routes } from 'react-router';
 import Nomatch from './app/Nomatch';
 import HomePublic from './app/public/home/Home';
 import useAuth from './hooks/useAuth';
-import Login from './app/public/login/Login';
 import Home  from './app/private/home/Home';
+import Login from './app/public/login/Login'; //login, viejo, no se que hace esto
+import LoginRegisterForm from './app/public/LoginRegisterForm/LoginRegisterForm'; //este es el login  +register que yo hice
 
 import CreatePostForm from './components/private/CreatePostForm/CreatePostForm';
 
@@ -44,7 +45,7 @@ export function App() {
 				<Route
 					path="/"
 					element={
-						isLogged ? <Home mode={darkMode} /> : <HomePublic />
+						isLogged ? <Home mode={darkMode} /> : <LoginRegisterForm />
 					}
 				/>
 				<Route path="/login" element={<Login login={loginUser} />} />
