+++ src/app/public/LoginRegisterForm/LoginRegisterForm.module.css
@@ -1,6 +1,6 @@
 form {
     background-color: #15202b;
-    padding: 250px;
+    padding: 50px;
     border-radius: 10px;
     color: white;
 }
