+++ src/app/public/LoginRegisterForm/LoginRegisterForm.tsx
@@ -1,0 +1,60 @@
+import { useState } from 'react';
+import './LoginRegisterForm.module.css';
+
+const LoginRegisterForm = () => {
+  const [loginEmail, setLoginEmail] = useState('');
+  const [loginPassword, setLoginPassword] = useState('');
+  const [registerEmail, setRegisterEmail] = useState('');
+  const [registerPassword, setRegisterPassword] = useState('');
+
+  const handleLogin = (e: React.FormEvent) => {
+    e.preventDefault();
+    console.log('Logging in with', loginEmail, loginPassword);
+  };
+
+  const handleRegister = (e: React.FormEvent) => {
+    e.preventDefault();
+    console.log('Registrando con', registerEmail, registerPassword);
+  };
+
+  return (
+    <div>
+      <form onSubmit={handleLogin}>
+        <h2>Iniciar sesión</h2>
+        <input
+          type="email"
+          placeholder="Correo electrónico"
+          value={loginEmail}
+          onChange={(e) => setLoginEmail(e.target.value)}
+        />
+        <input
+          type="password"
+          placeholder="Contraseña"
+          value={loginPassword}
+          onChange={(e) => setLoginPassword(e.target.value)}
+        />
+        <button type="submit">Iniciar sesión</button>
+      </form>
+
+      <form onSubmit={handleRegister}>
+      <h2>Registrarse</h2>
+      <input
+        type="email"
+        placeholder="Correo electrónico"
+        value={registerEmail}
+        onChange={(e) => setRegisterEmail(e.target.value)}
+      />
+      <input
+        type="password"
+        placeholder="Contraseña"
+        value={registerPassword}
+        onChange={(e) => setRegisterPassword(e.target.value)}
+      />
+      <button type="submit">Registrarse</button>
+      </form>
+    </div>
+  );
+};
+
+export default LoginRegisterForm;
+
