--- src/components/public/LoginForm/LoginForm.tsx
@@ -1,35 +1,0 @@
-import { useState } from 'react';
-import './LoginForm.module.css';
-
-const LoginForm = () => {
-  const [email, setEmail] = useState('');
-  const [password, setPassword] = useState('');
-
-  const handleLogin = (e: React.FormEvent) => {
-    e.preventDefault();
-    console.log('Logging in with', email, password);
-  };
-
-  return (
-    <form onSubmit={handleLogin}>
-      <h2>Iniciar sesión</h2>
-      <input
-        type="email"
-        placeholder="Correo electrónico"
-        value={email}
-        onChange={(e) => setEmail(e.target.value)}
-      />
-      <input
-        type="password"
-        placeholder="Contraseña"
-        value={password}
-        onChange={(e) => setPassword(e.target.value)}
-      />
-      <button type="submit">Iniciar sesión</button>
-    </form>
-
-  );
-};
-
-export default LoginForm;
-
