--- src/components/public/RegisterForm/RegisterForm.module.css
@@ -1,31 +1,0 @@
-form {
-    background-color: #15202b;
-    padding: 250px;
-    border-radius: 10px;
-    color: white;
-}
-  
-input {
-    display: block;
-    width: 100%;
-    margin-bottom: 15px;
-    padding: 10px;
-    background-color: #192734;
-    border: none;
-    border-radius: 5px;
-    color: white;
-}
-  
-button {
-    background-color: #1d9bf0;
-    color: white;
-    padding: 10px 20px;
-    border: none;
-    border-radius: 20px;
-    cursor: pointer;
-    font-weight: bold;
-}
-  
-button:hover {
-    background-color: #1a8cd8;
-}
