--- src/components/public/RegisterForm/RegisterForm.tsx
@@ -1,33 +1,0 @@
-import { useState } from 'react';
-import './RegisterForm.module.css';
-
-const RegisterForm = () => {
-  const [email, setEmail] = useState('');
-  const [password, setPassword] = useState('');
-
-  const handleRegister = (e: React.FormEvent) => {
-    e.preventDefault();
-    console.log('Registrando con', email, password);
-  };
-
-  return (
-    <form onSubmit={handleRegister}>
-      <h2>Registrarse</h2>
-      <input
-        type="email"
-        placeholder="Correo electrónico"
-        value={email}
-        onChange={(e) => setEmail(e.target.value)}
-      />
-      <input
-        type="password"
-        placeholder="Contraseña"
-        value={password}
-        onChange={(e) => setPassword(e.target.value)}
-      />
-      <button type="submit">Registrarse</button>
-    </form>
-  );
-};
-
-export default RegisterForm;
